FIX grayscale wraparound
We saw some weird behavior! When we added the numbers, if the total exceeded 255, it would wrap around to zero. The result was bright particles with dark centers, which caused subtle tracking errors. The solution is to take an average, and it is crucial to divide BEFORE adding. (I also wrote a comment about this.)